+++ Log.cs
@@ -10,11 +10,9 @@ public class Log
     public const string Prerequisites = "Prerequisites";
     public const string SavesLocation = "SaveLocation";
 
-    private static ILogger? _logger;
-
     public static ILogger Logger
     {
-        get => (_logger ?? Serilog.Log.Logger).ForContext("RemnantLogLibrary", "lib.remnant2.analyzer");
-        set => _logger = value;
+        get => (field ?? Serilog.Log.Logger).ForContext("RemnantLogLibrary", "lib.remnant2.analyzer");
+        set;
     }
 }